Move rflink base entity to separate module
