+++ homeassistant/components/rflink/cover.py
@@ -18,7 +18,6 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import RflinkCommand
 from .const import (
     CONF_ALIASES,
     CONF_DEVICE_DEFAULTS,
@@ -29,6 +28,7 @@ from .const import (
     CONF_SIGNAL_REPETITIONS,
     DEVICE_DEFAULTS_SCHEMA,
 )
+from .entity import RflinkCommand
 
 _LOGGER = logging.getLogger(__name__)
