+++ homeassistant/components/rflink/light.py
@@ -20,7 +20,6 @@ import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import SwitchableRflinkDevice
 from .const import (
     CONF_ALIASES,
     CONF_AUTOMATIC_ADD,
@@ -35,6 +34,7 @@ from .const import (
     EVENT_KEY_COMMAND,
     EVENT_KEY_ID,
 )
+from .entity import SwitchableRflinkDevice
 from .utils import brightness_to_rflink, rflink_to_brightness
 
 _LOGGER = logging.getLogger(__name__)
