+++ tests/components/rflink/test_switch.py
@@ -7,7 +7,7 @@ control of Rflink switch devices.
 
 import pytest
 
-from homeassistant.components.rflink import EVENT_BUTTON_PRESSED
+from homeassistant.components.rflink.entity import EVENT_BUTTON_PRESSED
 from homeassistant.const import (
     ATTR_ENTITY_ID,
     SERVICE_TURN_OFF,
